First attempt at generating a code object from the AST
TestInterp2 does not execute any Python code, but it generates the name
arrays (tuples) of a code object using the symbol table proved in
TestInterp1. Examples and reference results are generated by
code_testgen.py, and pasted in by hand. It works for globals and
positional arguments. The symbol table design caters keyword argument
and for closures, but it's not working yet.