Convert ASDL compiler to a Gradle plug-in
We add a refactored version of the ASDL compiler, wrapped as a Gradle
plug-in, alongside the existing compiler (perhaps now redundant). It
runs as a plug-in in the runtime subproject. It needs a little more work
to become a general-purpose plug-in like the ANTLR one.

We are not yet using the output in place of the checked-in
TreePython.java. Work is still needed to express dependency of the Java
compilation on this output.

The tools subproject has become an "included" project, mainly to
facilitate dependency ordering.