Improvements to the ASDL compiler and plug-in
This is driven by trying to use it to generate TreePython.java, and
although there are still problems with the template, this is a good
place to check-in.

Note that runtime:test fails on access to fields and the non-existence
of Num, which has gone from CPython's AST in Python.asdl so this is
still a broken build.