Get TestEx1.java compiling again
This involves some adaptation to the Python 3.8 AST, and along the way,
improvements to the ASDL compiler and default code generation. File
movements are in the interest of making the structure slightly less
deep.

The changes here are reflected in the chapter on "An Interpreter in Java
for the Python AST".

The build is still broken for examples after TestEx1 and that chapter.