Get TestEx[2-9].java compiling again
We separate the group files (templatres) for different phases of
development in the journey through the examples. Some of the added
features (in ExexNode.java.stg) should perhaps be part of the standard
template.

A few bugs are shaken out of the ASDL compiler and Gradle task in the
process.

TestInterp[1-5].java are still not compiling.