Fix compilation of TestInterp[1-5]: clean-running tests
In this change, adjustments are made to TestInterp[1-5] to match the
changes to the Python AST and the embedded examples are re-created with
the Python programs. The abstract base visitor is restored in the
template group file ExecNode.java.stg and used in the test programs.

The Gradle target test now runs cleanly.