Support specified base class and abstract visitor in ASDL compiler.
We do away with the specialised group files previously needed to support
evaluation of the AST, extending the built-in Java template to cover our
needs generically. The logic of the templates is re-worked, and support
in the compiler CodeTree, that controls when to use enums and what is
meant by a simple or built-in type.

runtime.gradle is adjusted to benefit from the new features.