Sequence access slots and some re-work of Slot and PyType
We painstakingly reproduce item get and set from the C API, while
implementing the slots they access. The signature enums are re-thought,
with the realisation that a rich set of signatures may be more
conveniently defined than previously.

There is some re-work of the "Type and Arithmetic" section of the
narrative to match.