Introduce Interpreter object and test an example callable
We set out to experiment with the call opcodes, but the strain this puts
on our understanding of stacks, name spaces, interpreter state and
thread state takes us to a new level architecturally, before ever we get
to proper functions.

The Interpreter object allows us to simplify test generation, and
associated changes invalidate some API, so the evo3 tests have to be
regenerated too.