Implement modules and functions
This is a big step! We define the structures critical for modules
defined in Java, and for exposing the functions within them using method
handles. There is now a builtins module with a few representative
functions in it. As a result, we can call len().

The apparatus of calls is revised to give the slot a strict signature,
since CPython assumes this, having to check it beforehand.