First implementation of PyFunction call
This change is largely the result of studying frame creation in CPython,
particularly ceval.c:: _PyEval_EvalCodeWithName() and the functions that
call it. The large function in C has been broken up into methods
attributed to PyFrame in the Java version, which is very different but
intends to preserve the logic. It is checked in as soon as one relevant
additional test has passed. A lot more test and fettling is required.