Define the form of generated code tests for evo4
The test structure is a small advance on evo3, the main innovation being
that when a result is an object appearing in __builtins__, such as a
built-in type or function, the reference result for comparison will be
found in the interp.builtinsModule. We add PyObjectTestEmitterEvo4 to
vsj2.srcgen and regenerate the test materials for evo4 using it.