A thorough analysis of slot definitions in CPython
These are updates to the narrative and architecture, checking beforehand
that restructuring slots for a simple alignment to the data model still
works in all the hard cases the CPython slotdefs can throw up.

The description is out of sync with the code now, and somewhat
inconsistent within itself, but I'd rather check this in, because of its
size, before going on to code.