Add member descriptors to the dictionary of the type
Also add a simple get test to PyTypeTest (more needed). The API of the
exposer changes as we think more precisely about the relationship of the
type and the class to expose.

A problem with circular dependency in the readying of types comes into
view but doesn't yet bite us.