Exposure of methods (partial implementation)
This change introduces PyMethodDescr and an interface intended as an
equivalent to the CPython vector call protocol. The exposure mechanism
is extended to methods defined in Java.

Little of this is integrated into the interpreter, but there is a test
to exercise it in isolation, and this is a useful way-point (with
passing tests) at which to commit the work.