Implement bound built-in methods
This re-factors PyJavaFunction so there is a base class PyJavaCallable
and subclasses for method and function. Tests are added for simple bound
methods. Complex cases (keywords and defaults) are not yet implemented,
nor is the vector call properly.

It is still not clear how to structure this so that we may take proper
advantage of call sites.