Clean up some deprecation warnings
These crop up where the approach to something has changed, so we know
the method implementing it should fall out of use, but finding and
correcting all the uses would be off-topic. This change goes back to fix
some of them. Often, there's a lot more to do than just change the
signature.