Experiments with class-specific binary operations
We try the idea of supplying and calling specific method signatures like
__add__(Double, PyFloat) as a way to make efficient call sites. Here is
an infrastructure intended to be general-purpose, but which so far only
works for limited examples of float.

In the course of this, we also clarify what we should mean by adopted,
accepted and operand Java classes, in relation to implementations, and
put in he infrastructure to support the concept. We can't test this
until we have more than one type.