Further support for String as str and some tests
This change adds code generation for PyUnicodeMethods and applies it to
generate comparison methods. The potential is there to generate unary
and binary operations, but these are few (and irregular) enough that we
prefer to hand-craft them.

Some deprecation warnings still occur where client code assumes only
PyUnicode can be str, but tests pass.