Re-work PyDict to wrap rather than inherit a Java map
In the plain object design, we must control the meaning of Java
hashCode() and equals() precisely in order to make keys equal in Python
act as the same key in a dict. There is a substantial test of this.

PyDict gains a member "map" and we use a Key object that wraps the
actual key. We add Abstract.hash() to the API.