Introduce ArgParser into MethodDef
Changes to code organisation, the main one being to introduce ArgPArser
to hold much of the information in a MethodDef. Some refactoring of
classes around PyMethodDescr to enable or benefit from this.

MethodDescriptor is defined as a common base for PyWrapperDescr and
PyMethodDescr, but after refactoring, this seems less useful.