Support more annotations in exposed methods and MethodDef
This change re-writes large parts of the processing of parameters on
exposed methods and the mechanism to call them. We introduce the notion
of a primary definition of an exposed method.

The code has many rough edges but is now much closer to the full
repertoire necessary to express exposed methods. Systematic tests are
provided to show that positional-only arguments are treated properly.

str.ljust is implemented to illustrate a method with a default argument.