Significantly re-work type and module exposure
A framework is provided for type and module exposure, and applied to
methods identified by annotations. Extensive signature processing is
implemented so that methods in these places have the full expressive
power of Python methods.

A module definition structure (ModuleDef) is generated similar to that
in CPython. The implementation of method definitions (MethodDef) is
improved and the way in which method descriptors are created.

Method definitions hold a parser that supports the call. This is a
different approach from CPython's.

The fundamental mechanisms are carefully tested, but tests do not cover
the full range of features sketched in. (It did not seem possible to
make this incrementally, and th change is already too big.)