Re-introduce get-set descriptors
We convert PyGetSetDescr from its VSJ 2 form to the new exposer and
plain objects pattern, supporting Getter, Setter and Deleter
annotations. There is a good test adapted from the one for Member, which
dabbles with attributes that may be native arrays internally and tuples
externally.

Still to do: specific support for PyGetSetDescr.Single, a proper test of
multiple accepted implementations, and to allow Getters and setters to
have their natural argument types (as exposed methods do).