Implement int division and modulus
These are evolved versions of the Jython 2 code, moved to PyLongMethods.
We generate wrappers that use thanks to additions to PyLong.py. These
operations are not simply invocations of the Java counterparts but quite
complex, so there is a test, to avoid looking foolish when we can
eventually run Python regression tests.