Move optimisation to signature into PyMethodDescr
Prior to this change, we depended on sub-classing MethodDef to express
short-cuts available when a method signature has a simple form (a fixed
number of positional parameters, say). We now sub-class directly in
PyMethodDescr. The motivations for this are directness and the potential
to optimise call sites by the class of the callable. The choice of
sub-class moves from the Exposer to static PyMethodDescr.fromParser.

We cannot un-pick the logic in MethodDef until the same changes are
complete in PyJavaMethod, at which point MethodDef can probably be
removed entirely.