Bind the handle in a PyJavaMethod to self
Rather than make self an argument internally to the handle invocation.
This simplifies the call to methods exposed by a Python module defined
in Java. MethodSignature provides the binding logic. Some thinking
errors are corrected in the process.

A sub-class of PyJavaMthod is added, and its tests, to exploit the
special case of MethodSignature.O1, a function with one argument.