Implement or re-work concatenation in str, bytes and tuple
We use a common approach to these three, supported by PyObjectUtil and
the PySequenceInterface, which now has a specialisation to primitive
int. The interface is expanded to provide Stream (and IntStream) access
to elements, and this is used to implement concatenation.

There is a unit test of getItem, concat and repeat, which identified
some corner cases where lone surrogates combine accidentally. It is now
clear what semantics to give to PyUnicode and String where non-BMP
characters are concerned.