Port formating apparatus from Jython 2
We introduce FieldsNameIterator, MarkupIterator, the specification
parser and the formatters for str, int and float. Then we apply these in
str, int and float. The aim is to convert these to the new core and push
them back into Jython 3. We also port from Jython 2 (and greatly
refactor) tests that are fairly thorough for int and the parser, but do
not much exercise other types.