Read example code objects
We add marshal codecs for bytes, code and float so that we can read the
example pyc files. Test are added for marshal.load of float objects, and
a test that exercises reading code. objects from pyc files.

We do not yet try to do anything with the code objects.