Obtain frame built-ins via Interpreter
Each frame now has a non-null builtins member which is obtained via that
frame's interpreter. This implements an important architectural choice
in Jython.

The test of this is accessing bool.__add__.__name__ in the test material
attr_access_builtin.py. We implement the exposure of inherited fields
(__name__ from Descriptor) to make this tidy.