Implement call to functions defined in Python
In this change we add a __call__ slot method to PyFunction, and exercise
it with a new "Python example" test, for which we add some new opcodes
supporting fast local variables. Along the way we fix a bug in ArgParser
and improve the tool that compiles examples to disassemble all code
objects even if running the code fails.

We have not yet addressed cells or LOAD_GLOBAL.