Implement ThreadState and the per-thread frame stack
We ground the previously "loose" frame stack in a ThreadState in a
ThreadLocal variable. This completes the architectural pattern we think
Interpreter, PyFunction and PyCode should have to platform threads.

We implement builtins.exec as a way of developing the API but have yet
to devise a test for it.