Support Python 3.11 in name only
While the code still compiles, we rename anything with 38 in the name to
have 311 instead. Tests are disabled where they now fail, which is
anything needing to de-marshal a code object.

Renaming affects a fair number of files so this is to get that noise out
of the way before the interesting bits.