Implement the type system in outline
This is quite a sophisticated collection of classes reflecting the ideas
in the narrative concerning types, and also ideas of encapsulation that
minimises the exposure of internals as API.

A major concern of this installment has been robust initialisation of
the type system.

The type system involves multiple interconnected singleton objects. This
makes it difficult to get right and difficult to test. It may not be the
right approach.