Robust static initialisation of the type system
We make a comprehensive attempt to guarantee proper creation of a
static, singleton, type system in a multi-threaded application. There is
a test to prove we succeeded.

The type system does not yet create descriptors, but it could, and the
test will fail later if that cannot safely be introduced.