Thread-safe bootstrap pattern
We implement a pattern for bootstrap types that appears solid and is
much simpler than the idea of a hidden implementation class. So far, we
only test it with a simple PyFloat.

We also prepare for a new approach to subclasses defined in Python, by
removing the ExtensionPoint interface and related code.