Specify subclass representation
This change begins to address subclasses of built-in types in a properly
general way. It appears necessary to generate new classes in a stylised
way so that arbitrary combinations of Java superclasses and interfaces
can be catered for. We create a specification object and show we can
generate simple subclasses of object and float.