Create some exceptions through __new__ and __init__
This works for BaseException and exceptions with the same Java
representation class. It doesn't work yet for those with a subclass
representation (and no __new__ of their own), which is required.

This work fills important gaps in the type system, as a result of
solving this problem, although special methods and their caches are
still an inconsistent mess.