Create instances of subclasses of exception reflectively
We use the constructor reflectively to create instances of the
subclasses as there is no __new__ for the subclass. (We're following
CPython closely.) This allows us to explore through tests the subclass
exceptions that take keyword arguments to __init__.

This approach requires that the constructor be public. That may be
acceptable here but is undesirable and may be not feasible in general.