Move objects to core and flatten package tree (#10)
The primary motivation is to place object definitions (PyFloat etc.) into core as is familiar from Jython 2.

In recent work, we set out to have a runtime package with just a few bootstrap types in it, and the others in core. Most of the types we have defined turned out to be bootstrap types. This means the core types all ended up in runtime, but this will be an obstacle to transitioning the work into Jython proper. Hence a big renaming is called for. We take the opportunity to rationalise package names too.