More Javadoc cleanup
This really is pretty grubby ... a lot oflinting may be silenced by
making things private. This is the right answer for special methods. Not
so clear that Python API should not be callable from Java, in which case
writing documentation is the answer.

In the process we improve our understanding of some methods to do with
method signatures.